chore: refactored obs module structure
