+++ .github/workflows/develop.yaml
@@ -28,6 +28,6 @@ jobs:
         distribution: 'temurin'
         cache: maven
     - name: Build with Maven
-      run: cd online-banking-service && mvn clean install
+      run: mvn clean install
 
 #Add pmd check
