+++ obs/obs-rest-server/pom.xml
@@ -23,4 +23,28 @@
         <ruleset.basedir>../..</ruleset.basedir>
     </properties>
 
+    <dependencies>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-web</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.adorsys.webank</groupId>
+            <artifactId>obs-service-api</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.adorsys.webank</groupId>
+            <artifactId>obs-rest-api</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <!-- Other existing dependencies -->
+    </dependencies>
+
+
 </project>
