+++ obs/obs-service-api/src/main/java/com/adorsys/webank/obs/OBSServiceApi.java
@@ -1,0 +1,6 @@
+package com.adorsys.webank.obs;
+
+public interface OBSServiceApi {
+    String getMessage();
+}
+
