+++ obs/obs-service-impl/src/main/java/com/adorsys/webank/obs/OBSServiceImpl.java
@@ -1,0 +1,12 @@
+package com.adorsys.webank.obs;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class OBSServiceImpl implements OBSServiceApi {
+
+    @Override
+    public String getMessage() {
+        return "Hello from the OBS Service!";
+    }
+}
