+++ online-banking-app/src/main/java/com/adorsys/webank/OnlineBankingApplication.java
@@ -1,0 +1,13 @@
+package com.adorsys.webank.obs;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OnlineBankingApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(OnlineBankingApplication.class, args);
+	}
+
+}
