+++ online-banking-app/src/test/java/com/adorsys/webank/OnlineBankingApplicationTests.java
@@ -1,0 +1,14 @@
+package com.adorsys.webank.obs;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest(classes = OnlineBankingApplication.class) // Specify main class
+public class OnlineBankingApplicationTests {
+
+	@Test
+	public void contextLoads() {
+		// This test intentionally does nothing.
+		// The @SpringBootTest annotation simulates application startup.
+	}
+}
