--- online-banking-service/src/main/java/com/adorsys/webank/obs/OnlineBankingServiceApplication.java
@@ -1,13 +1,0 @@
-package com.adorsys.webank.obs;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OnlineBankingServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OnlineBankingServiceApplication.class, args);
-	}
-
-}
