--- online-banking-service/src/main/java/com/adorsys/webank/obs/controller/OBSRestController.java
@@ -1,13 +1,0 @@
-package com.adorsys.webank.obs.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class OBSRestController {
-
-    @GetMapping("/message")
-    public String getMessage() {
-        return "Hello from OBService!";
-    }
-}
