--- online-banking-service/src/main/java/com/adorsys/webank/obs/service/api/OBServiceAPI.java
@@ -1,6 +1,0 @@
-package com.adorsys.webank.obs.service.api;
-
-
-public interface OBServiceAPI {
-    String getMessage();
-}
