--- online-banking-service/src/test/java/com/adorsys/webank/obs/OnlineBankingServiceApplicationTests.java
@@ -1,14 +1,0 @@
-package com.adorsys.webank.obs;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest(classes = OnlineBankingServiceApplication.class) // Specify main class
-public class OnlineBankingServiceApplicationTests {
-
-	@Test
-	public void contextLoads() {
-		// This test intentionally does nothing.
-		// The @SpringBootTest annotation simulates application startup.
-	}
-}
