--- online-banking-service/src/test/java/com/adorsys/webank/obs/controller/OBSRestControllerTest.java
@@ -1,23 +1,0 @@
-package com.adorsys.webank.obs.controller;
-
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
-
-@WebMvcTest(OBSRestController.class)
-public class OBSRestControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Test
-    public void testGetMessage() throws Exception {
-        mockMvc.perform(MockMvcRequestBuilders.get("/message"))
-                .andExpect(MockMvcResultMatchers.status().isOk())
-                .andExpect(MockMvcResultMatchers.content().string("Hello from OBService!"));
-    }
-}
